chore: update to new hammer version fix: sync issues with fluid storage, energy bar rendering

+++ src/src/main/java/com/github/mixinors/astromine/common/config/section/WorldConfigSection.java
@@ -13,7 +13,7 @@ public class WorldConfigSection {
 	public OresConfigSection ores = new OresConfigSection();
 	
 	@Comment("Gravity level in Space")
-	public double spaceGravity = 0.01D;
+	public float spaceGravity = 0.01F;
 	
 	@Comment("Gravity for non-Astromine dimensions")
 	public double gravity = 0.08D;
